remove event polling from main cpu loop
since we are not in a gui environment, there is no need
to poll for events.  So I took the polling out of the
main cpu loop and instead added it to the keyboard read
code.

Note that we now have 2 keyboard queues - the scantran one
that I wrote and the one used in xtrs.   We could maybe get
rid of one.  But having 2 is convenient because it makes it
easy to handle synthetic shifts