First pass at new realtime logic
New realtime logic that essentially lets the z80 instructions
run as fast as they can until an IO operation occurs and then
forces them to sync.   But we don't want "emulator" time to ever
run too far ahead of real time so we force a sync every now and then
regardless.

Anecdotally, it seems like the emulation speed sometimes runs faster
than others, so there may be a bug in there somewhere

Finally, I fixed a bug where we didn't emulate the way the model 1
generates A6 of the character rom, which caused some of the level 2
characters to look wrong.